feat(MenuItem): separate reusable MenuItem
