+++ src/components/views/Sheets/Sheets.stories.tsx
@@ -4,10 +4,7 @@ import { ComponentMeta, ComponentStory } from "@storybook/react"
 import { userEvent, within } from "@storybook/testing-library"
 import { IconButton } from "../../controls/buttons/IconButton/IconButton"
 import { PrimaryButton } from "../../controls/buttons/PrimaryButton/PrimaryButton"
-import { Menu } from "../../menus/Menu/Menu"
-import { MenuContent } from "../../menus/Menu/MenuContent"
-import { MenuItem } from "../../menus/Menu/MenuItem"
-import { MenuTrigger } from "../../menus/Menu/MenuTrigger"
+import * as Menu from "../../menus/Menu"
 import { Text } from "../../typography/Text/Text"
 import { Sheets } from "./Sheets"
 
@@ -115,15 +112,15 @@ const ContextMenuTemplate: ComponentStory<typeof Sheets> = (args) => {
   const { isOpen, handlers } = useDisclosure(true)
   return (
     <>
-      <Menu>
-        <MenuTrigger>
+      <Menu.Root>
+        <Menu.Trigger>
           <IconButton aria-label="See options" icon="ellipsis" />
-        </MenuTrigger>
-        <MenuContent>
-          <MenuItem onClick={handlers.open}>Option 1</MenuItem>
-          <MenuItem onClick={handlers.open}>Option 2</MenuItem>
-        </MenuContent>
-      </Menu>
+        </Menu.Trigger>
+        <Menu.Content>
+          <Menu.Item onClick={handlers.open}>Option 1</Menu.Item>
+          <Menu.Item onClick={handlers.open}>Option 2</Menu.Item>
+        </Menu.Content>
+      </Menu.Root>
       <Sheets
         {...args}
         closeHandler={handlers.close}
